clean code step 1 (#339)

+++ apps/api/src/controllers/joconde.js
@@ -1,3 +1,4 @@
+// TODO: control authorization (can use create, update or delete based on role, group and museo?).
 const express = require("express");
 const router = express.Router();
 const multer = require("multer");
@@ -8,29 +9,19 @@ const upload = multer({ dest: "uploads/" });
 const Joconde = require("../models/joconde");
 const Museo = require("../models/museo");
 const { capture } = require("./../sentry.js");
-
 const { uploadFile, deleteFile, formattedNow, checkESIndex, updateNotice } = require("./utils");
-
-function transformBeforeUpdate(notice) {
-  if (notice.IMG !== undefined) {
-    notice.CONTIENT_IMAGE = notice.IMG ? "oui" : "non";
-  }
-  notice.DMAJ = formattedNow();
-}
 
 async function checkJoconde(notice) {
   const errors = [];
   try {
-    //Check contact
+    // Check contact
     if (!notice.CONTACT) {
       errors.push("Le champ CONTACT ne doit pas être vide");
     }
-
-    //Joconde
+    // Joconde
     if (!notice.TICO && !notice.TITR) {
       errors.push("Cette notice devrait avoir un TICO ou un TITR");
     }
-
     for (let i = 0; i < IMG.length; i++) {
       try {
         await rp.get(PREFIX_IMAGE + IMG[i]);
@@ -41,8 +32,14 @@ async function checkJoconde(notice) {
   } catch (e) {
     capture(e);
   }
-
   return errors;
+}
+
+function transformBeforeUpdate(notice) {
+  if (notice.IMG !== undefined) {
+    notice.CONTIENT_IMAGE = notice.IMG ? "oui" : "non";
+  }
+  notice.DMAJ = formattedNow();
 }
 
 function transformBeforeCreate(notice) {
@@ -68,6 +65,7 @@ function transformBeforeCreate(notice) {
   });
 }
 
+// Update a notice by ref
 router.put(
   "/:ref",
   passport.authenticate("jwt", { session: false }),
@@ -78,38 +76,39 @@ router.put(
 
     try {
       const prevNotice = await Joconde.findOne({ REF: ref });
-
       const arr = [];
 
+      // Delete previous images if not present anymore (only if the actually is an IMG field).
       if (notice.IMG !== undefined) {
         for (let i = 0; i < prevNotice.IMG.length; i++) {
           if (!(notice.IMG || []).includes(prevNotice.IMG[i])) {
-            arr.push(deleteFile(prevNotice.IMG[i]));
+            // Security: no need to escape filename, it comes from database.
+            if (prevNotice.IMG[i]) {
+              arr.push(deleteFile(prevNotice.IMG[i]));
+            }
           }
         }
       }
 
+      // Upload all files.
       for (let i = 0; i < req.files.length; i++) {
         const f = req.files[i];
-        console.log(
-          "uploadFile",
-          `joconde/${filenamify(notice.REF)}/${filenamify(f.originalname)}`
-        );
-        arr.push(uploadFile(`joconde/${filenamify(notice.REF)}/${filenamify(f.originalname)}`, f));
+        const path = `joconde/${filenamify(notice.REF)}/${filenamify(f.originalname)}`;
+        arr.push(uploadFile(path, f));
       }
 
-      // Update IMPORT ID
+      // Update IMPORT ID (this code is unclear…)
       if (notice.POP_IMPORT.length) {
         const id = notice.POP_IMPORT[0];
         delete notice.POP_IMPORT;
         notice.$push = { POP_IMPORT: mongoose.Types.ObjectId(id) };
       }
 
+      // Prepare and update notice.
       transformBeforeUpdate(notice);
-
-      // Update Notice
       arr.push(updateNotice(Joconde, ref, notice));
 
+      // Consume promises and send sucessful result.
       await Promise.all(arr);
       res.sendStatus(200);
     } catch (e) {
@@ -119,6 +118,7 @@ router.put(
   }
 );
 
+// Create a new notice.
 router.post(
   "/",
   passport.authenticate("jwt", { session: false }),
@@ -126,21 +126,19 @@ router.post(
   async (req, res) => {
     try {
       const notice = JSON.parse(req.body.notice);
-
       const arr = [];
-      for (var i = 0; i < req.files.length; i++) {
-        arr.push(uploadFile(`joconde/${notice.REF}/${req.files[i].originalname}`, req.files[i]));
+
+      // Upload all files (should this be done after creating notice?)
+      for (let i = 0; i < req.files.length; i++) {
+        const path = `joconde/${filenamify(notice.REF)}/${filenamify(req.files[i].originalname)}`;
+        arr.push(uploadFile(path, req.files[i]));
       }
 
+      // Transform and create.
       await transformBeforeCreate(notice);
-
       const obj = new Joconde(notice);
-
-      // Send error if obj is not well sync with ES
       checkESIndex(obj);
-
       arr.push(obj.save());
-
       await Promise.all(arr);
       res.send({ success: true, msg: "OK" });
     } catch (e) {
@@ -150,6 +148,8 @@ router.post(
   }
 );
 
+// Get notices by offset limit. Not sure it's still in use.
+// TODO: check if it's in use.
 router.get("/", (req, res) => {
   const offset = parseInt(req.query.offset) || 0;
   const limit = parseInt(req.query.limit) || 20;
@@ -158,6 +158,7 @@ router.get("/", (req, res) => {
   });
 });
 
+// Get one notice by ref.
 router.get("/:ref", (req, res) => {
   const ref = req.params.ref;
   Joconde.findOne({ REF: ref }, (err, notice) => {
@@ -173,24 +174,24 @@ router.get("/:ref", (req, res) => {
   });
 });
 
+// Delete one notice.
 router.delete("/:ref", passport.authenticate("jwt", { session: false }), async (req, res) => {
   try {
     const ref = req.params.ref;
     const doc = await Joconde.findOne({ REF: ref });
     if (!doc) {
       return res.status(500).send({
-        error: `Je ne trouve pas la notice joconde ${ref} à supprimer`
+        error: `Impossible de trouver la notice joconde ${ref} à supprimer.`
       });
     }
-    const arr = doc.IMG.map(f => deleteFile(f));
+    // remove all images and the document itself.
+    const arr = doc.IMG.filter(i => i).map(f => deleteFile(f));
     arr.push(doc.remove());
     await Promise.all(arr);
     return res.status(200).send({});
   } catch (error) {
     capture(error);
-    return res.status(500).send({
-      error
-    });
+    return res.status(500).send({ error });
   }
 });
